test(agent): isolate replay complexity regression
